scripts/diffusion-requirements.txt
# diffusion부분을 git clone 해온 후에 diffusion/scripts/inference.py 안에 model.cuda()로 되어있는 부분을
# device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# model.to(device)
# 로 바꿔줘야한다. 안 그러면 cpu만 있는 컴퓨터에서는 오류가 발생한다.
# 추가로
# c = model.get_learned_conditioning(ref_tensor.to(torch.float16)) 인 부분을
# c = model.get_learned_conditioning(ref_tensor.to(torch.float32)) 로 바꿔줘야한다.
# 왜냐하면 PyTorch의 CPU 모드는 float32 연산만 지원하기 때문이다.
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==0.4.3
altair==5.4.1
antlr4-python3-runtime==4.8
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bezier==2023.7.28
bleach==6.1.0
blinker==1.8.2
boto3==1.36.26
botocore==1.36.26
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.1
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
-e git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1#egg=clip
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.32.2
einops==0.3.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.2.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==8.5.0
importlib-resources==6.4.5
invisible-watermark==0.2.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kornia==0.6.0
lazy-loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
multidict==6.1.0
narwhals==1.25.2
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.3.2
notebook-shim==0.2.4
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.1.2.30
opencv-python-headless==4.11.0.86
overrides==7.7.0
packaging==24.2
# Editable install with no version control (Paint-by-Example==0.0.1)
-e c:\stable-diffusion
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
pudb==2019.2
pure-eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydeck==0.9.1
pyDeprecate==0.3.1
pygments==2.19.1
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytorch-lightning==1.4.2
pytz==2025.1
PyWavelets==1.4.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.35.1
regex==2024.11.6
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scikit-image==0.20.0
scipy==1.9.1
Send2Trash==1.8.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.40.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.18.1
test-tube==0.7.5
tifffile==2023.7.10
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.2.1
torch==1.11.0
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.12.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.19.2
types-python-dateutil==2.9.0.20241206
typing-extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
urwid==2.6.16
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
widgetsnbextension==4.0.13
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
yarl==1.15.2
zipp==3.20.2
